jira: Plug update_field() into Issue class when we return it
This replaces the set_field() call with the logic we use to update fields
from the command line.  This enables callers to dynamically things
utilizing the input field resolution and type resolution/translation we
already had.

Callers who want to do it themselves can still call issue.update().

  issue = jirate_obj.issue('MYKEY-123')
  issue.update_field('fixed_in_build', 'my-rpm-1.2.3-1')
  issue.update_field('labels', 'mylabel', 'add')
