+++ jirate/jboard.py
@@ -34,20 +34,53 @@ def _resolve_field(obj, field_name):
     return obj._jirate.field(obj, field_name)
 
 
-def _set_field(obj, field_alias, value):
-    field_name = obj._jirate._field(obj, field_alias)
-    return obj.update({field_name: value})
+# Intelligent field update for Issue
+def _update_field(issue, field_name_human, value_human, operation='set', fields=None):
+    if not fields:
+        # TODO use native python-jira issue.fields instead of raw json
+        # (Except operations are not captured, which we need)
+        fields = issue._jirate.fields(issue.key)
+    if isinstance(value_human, list):
+        value_human = ' '.join(value_human)
+    else:
+        value_human = str(value_human)
+
+    # TODO multi-field sets?
+    field_args = {field_name_human: value_human}
+    output_args = transmogrify_input(fields, **field_args)
+
+    if not output_args:
+        raise AttributeError(f'No field like \'{field_name_human}\' in {issue.key}')
+
+    # Set up for the rest
+    field_ids = [key for key in output_args.keys()]
+    field_id = field_ids[0]
+    field = fields[field_id]
+    send_val = output_args[field_id]
+
+    ops = field['operations']
+    if operation not in ops:
+        raise ValueError(f'Cannot perform \'{operation}\' on \'{field_name_human}\' of {issue.key}; try: {ops}')
+
+    # Add and remove use a different format than 'set'.
+    # There's also 'modify', but ... that one's even more complicated.
+    if operation in ['add', 'remove']:
+        update_args = {field_id: [{operation: val} for val in send_val]}
+    else:
+        update_args = {field_id: [{operation: send_val}]}
+
+    return issue.update(**update_args)
 
 
 def _resolve_field_setup(jirate_obj, issue_obj):
     # Do NOT trample future python JIRA objects' field function.
     if hasattr(issue_obj, 'field'):
-        raise Exception('API BREAK: \'field\' is now part of jira.resources.Issue. Please file a bug!')
-    if hasattr(issue_obj, 'set_field'):
-        raise Exception('API BREAK: \'set_field\' is now part of jira.resources.Issue. Please file a bug!')
+        raise Exception('API BREAK: \'field\' is now part of jira.resources.Issue. Please file a bug against Jirate!')
+    if hasattr(issue_obj, 'update_field'):
+        raise Exception('API BREAK: \'update_field\' is now part of jira.resources.Issue. Please file a bug against Jirate!')
     issue_obj._jirate = jirate_obj
     issue_obj.field = types.MethodType(_resolve_field, issue_obj)
-    issue_obj.set_field = types.MethodType(_set_field, issue_obj)
+    issue_obj.update_field = types.MethodType(_update_field, issue_obj)
 
 
 def _check_fields(issue, name):
@@ -775,7 +808,7 @@ class JiraProject(Jirate):
 
     def _index_issue(self, issue):
         if issue.key not in self._config['issue_map']:
-            if not getattr(issue, '_jirate'):
+            if not hasattr(issue, '_jirate'):
                 _resolve_field_setup(self, issue)
             self._config['issue_map'][issue.key] = issue
